requirements.txt
# Django and Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
django-environ==0.11.2

# Authentication & Security
django-ratelimit==4.1.0
django-oauth-toolkit==1.7.1

# Caching
redis==5.0.1
django-redis==5.4.0

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# File Handling
Pillow==10.1.0
openpyxl==3.1.2
pandas==2.1.3

# Email
django-anymail==10.1
sendgrid==6.10.0

# API Documentation
drf-yasg==1.21.7
django-rest-swagger==2.2.0

# Monitoring & Logging
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.6.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0

# Additional Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
urllib3==2.1.0
